Revert "Revert "Add score per player and score reset button""
