+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## [1.2.0] - 2023-06-21
+
+### Added
+- Score per player.
+- Score reset button.
