+++ README.md
@@ -1,5 +1,11 @@
-# Tic Tac Toe
-It's a mobile video game for 2 players made in react native.
+# Tic-Tac-Toe
+It's a mobile video game made in react native.
 
 # Home
-![alt text](https://raw.githubusercontent.com/johnreyes96/udemy-tictactoe-app-using-react-native/release/1.1.1/assets/home.jpeg)
+![alt text](https://raw.githubusercontent.com/johnreyes96/udemy-tictactoe-app-using-react-native/0a596bef70a9099952eaa0aec0922b3e9f20a45a/assets/home.jpeg)
+
+## Usage
+It is a game for 2 people where the first player to make 3 in a row will win, there is a button to reset the score.
+
+## License
+Apache-2.0
